requirements.txt
streamlit==1.41.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
PyPDF2==3.0.1
nltk==3.9.1
joblib==1.4.2
matplotlib==3.9.2
seaborn==0.13.2
sentence-transformers==3.3.1
google-generativeai==0.8.4
